# lmdb: Return dedicated ErrNotFound from Txn.Get and Cursor.Get