This avoids allocation of a new Error every time we look up a key that does not exist, saving two allocations per lookup.

I have a use case where most lookups end up not existing in the database, where this comes in handy. I considered changing operrno at first, unfortunately mdb_dbi_open returns MDB_NOTFOUND as well. Returning ErrNotFound in that case would have been too confusing I think.

Not sure I got all the documentation changes right, lmk what needs fixing.