@bmatsuo 2016-03-31
This is an interesting proposal. Though I feel compelled to reject it as it is. It is not technically backwards compatible. Someone could potentially be counting on the value being wrapped. I don't really want to wager on the chance of that. 

Wrapping errno values with OpErrno was also a very deliberate choice. Before, when errno values were used directly, I found that passing errors up the call stack made tracing down a random MDB_NOT_FOUND or EINVAL to be extremely painful. While MDB_NOT_FOUND is special in the case of Txn.Get/Cursor.Get (though not really in mdb_dbi_open, as you point out), I'm still skeptical of the cost vs benefit in this change.

However, I am more open to introducing a new API that returns an Errno value instead of error (an interface, for which avoiding an allocation is actually impossible).

```
func (txn *Txn) GetErrno(dbi DBI, key []byte) ([]byte, []byte, Errno)
func (cur *Cursor) GetErrno(key []byte, val []byte, op uint) ([]byte, []byte, Errno)
```

This would be compatible, and I think the explicit nature of it is kind of nice. What do you think? I probably won't move on this today. So don't worry about updating ASAP.

It would also be helpful if you had some benchmark numbers about what kind of slowdown you actually would see in an application with/without this change. If you could come up with those numbers I would appreciate it.