@bmatsuo 2016-03-31
Oh yea. Did you actually see the number of allocations go down in that benchmark you wrote? Because, like I mentioned, I think a value type (non-pointer) will actually get allocated onto the heap when you put turn it into an interface value.

Sorry for not putting that together sooner.