@lmb 2016-03-31
Hmm, I see your point.

The allocations come from allocating a new OpError every time we get
MDB_NOTFOUND. So yeah, allocations drop. Runtime isn't my main concern, I
need to prevent churning the GC. Not sure I follow what you wrote re value
types being allocated in the heap.

Tbh, returning MDB_NOTFOUND is a hack that is necessary because C does not
have multiple return values. The natural interface in go would be Get
returning nil, nil if the key does not exist.

How about this: I turn ErrNotFound into an OpError with the appropriate
Errno, but a generic op?