@bmatsuo 2016-03-31
Here is what I mean about non-pointers being used as interface values:

First, just think about it. An interface has to contain a pointer to the contained value, because the contained value can have any size and it is not known at compile time. So if I return a 100 byte struct that implements `error` from a function, where is the interface's pointer pointing?

Now see this example: http://play.golang.org/p/9mtA2PxU4r

And this benchmark: http://play.golang.org/p/HtebvZSuKX

Running the benchmark shows that there is one alloc/op

```
BenchmarkFailThingy-4	30000000	        40.4 ns/op	       8 B/op	       1 allocs/op
```

And I would ask you again, are you **sure** that allocations dropped? It may be that the compiler is smart enough to know how much space is needed on the stack. It seemed with some versions of my example that was the case (I'm sure you notice the weird `if` statement -- it's there for a reason).

> How about this: I turn ErrNotFound into an OpError with the appropriate Errno, but a generic op?

That seems more technically compatible. Again, there is some (smaller) possibility of people checking Op. It also seems like a fairly unsightly hack to me. It's the kind of thing that I think real evidence is needed for.

And I'm sorry if I gave you the impression that I was talking about start-to-end runtime of a benchmark. I am talking about any valid measurement of performance in a practical application. That may be total throughput, wall clock time, or response time. Any such demonstration would be very helpful in deciding to do something less that fully compatible, though honestly the chances are not great in any case.

I don't like the bloat from additional functions (especially with the prospect of additional functions for querying using interface values -- #11). But I do know people who are seriously using this project in a commercial space. I don't want to do anything that would adversely effect those users.

Granted, I am using semver for this project. But I was not intending on breaking the API in the near future.