@bmatsuo 2016-04-01
I ran the benchmarks and the allocs/op do drop. FWIW they drop in all the Get benchmarks as they all seem to encounter MDB_NOTFOUND. It is a little suspicious actually. But I don't remember if they are supposed to do that and I don't have the time for a full investigation today.

Optimizations regarding non-pointer interfaces seem perhaps more sophisticated then I believed.

It might be ok to have transactions/cursors reuse an error stored in their structs. Then repeated misses might not be so bad. I don't know.

This whole situation deserves more thought and we can't proceed too quickly. Please keep thinking about alternative APIs that are idiomatic and make sense, or otherwise a way to mask things well that is more completely compatible.

I do appreciate the work you have put in. Thank you.