@lmb 2016-04-01
I reworked the PR to split out the benchmark to make before / after comparisons easier.

Some background on my workload: clients query LMDB through the memcached protocol. Most queries return a negative answer, think whitelisting / blacklisting. I took some real-life traffic, and replayed it against an empty LMDB (this is not 100% accurate obviously).

**Without this PR**
<img width="500" alt="without-pr" src="https://cloud.githubusercontent.com/assets/368392/14204365/cbfa28e0-f7fb-11e5-910c-04b073c06ec3.png">

**With this PR**
<img width="500" alt="with-pr" src="https://cloud.githubusercontent.com/assets/368392/14204399/f3d492c4-f7fb-11e5-8f91-d4be30918927.png">

**With this PR and the char* work**
<img width="500" alt="with-pr-and-charstar" src="https://cloud.githubusercontent.com/assets/368392/14204412/14a1bbf8-f7fc-11e5-8512-2b807b5e088c.png">

On the synthethic benchmarks:

```
$ benchcmp baseline.txt new.txt
benchmark                             old ns/op     new ns/op     delta
BenchmarkTxn_Get_missing_raw-4        507           408           -19.53%
BenchmarkCursor_Get_missing_raw-4     546           441           -19.23%

benchmark                             old allocs     new allocs     delta
BenchmarkTxn_Get_missing_raw-4        4              2              -50.00%
BenchmarkCursor_Get_missing_raw-4     4              2              -50.00%

benchmark                             old bytes     new bytes     delta
BenchmarkTxn_Get_missing_raw-4        84            48            -42.86%
BenchmarkCursor_Get_missing_raw-4     84            48            -42.86%
```